Move `surface_configure` out of `Global`
